+++ src/hooks/useGetMyProducts.js
@@ -1,0 +1,29 @@
+import useAxiosInstance from "@hooks/useAxiosInstance";
+import { useQuery } from "@tanstack/react-query";
+
+export const useGetMyProducts = (userId, select) => {
+  const axios = useAxiosInstance();
+
+  return useQuery({
+    queryKey: ["products", userId],
+    queryFn: () => {
+      return axios.get("/products", { params: { seller_id: userId } });
+    },
+    select: res => {
+      return select ? select(res.data.item) : res.data.item;
+    },
+    staleTime: 1000 * 10,
+  });
+};
+
+export const useMyProductsFilter = (userId, states) => {
+  return useGetMyProducts(userId, data => {
+    if (states.length === 0) {
+      return data;
+    } else {
+      return data.filter(product => {
+        return states.some(state => state === product.extra.state);
+      });
+    }
+  });
+};
